Refactor: Combine path_utils and url_utils into utils.py
Combines the functionality of `path_utils.py` and `url_utils.py` into a single file named `utils.py` within `plugin/framework/`.

This refactoring removes the two separate utility modules and consolidates them to simplify imports and logical grouping of basic string/path/URL utilities.

All references across the codebase to `plugin.framework.path_utils` and `plugin.framework.url_utils` have been updated to point to `plugin.framework.utils`.
