+++ plugin/tests/test_image_status_callback.py
@@ -1,6 +1,6 @@
 
 import sys
-from plugin.framework.path_utils import get_plugin_dir
+from plugin.framework.utils import get_plugin_dir
 import os
 import unittest
 from unittest.mock import MagicMock, patch
