+++ plugin/tests/test_search_web.py
@@ -20,7 +20,7 @@ You can pass a custom --max-tokens if desired, and override the model with --mod
 import argparse
 import json
 import os
-from plugin.framework.path_utils import get_plugin_dir
+from plugin.framework.utils import get_plugin_dir
 import sys
 import time
 from typing import Any
