+++ plugin/tests/uno/test_chat_model_logic.py
@@ -1,5 +1,5 @@
 import sys
-from plugin.framework.path_utils import get_plugin_dir
+from plugin.framework.utils import get_plugin_dir
 import os
 import unittest
 from unittest.mock import MagicMock, patch
